+++ src/lib/account_update.ts
@@ -1934,7 +1934,7 @@ function addMissingSignatures(
     let signature = signFieldElement(
       fullCommitment,
       privateKey.toBigInt(),
-      'testnet'
+      Mina.getNetworkId()
     );
     return { body, authorization: Signature.toBase58(signature) };
   }
@@ -1967,7 +1967,7 @@ function addMissingSignatures(
     let signature = signFieldElement(
       transactionCommitment,
       privateKey.toBigInt(),
-      'testnet'
+      Mina.getNetworkId()
     );
     Authorization.setSignature(accountUpdate, Signature.toBase58(signature));
     return accountUpdate as AccountUpdate & { lazyAuthorization: undefined };