+++ src/lib/mina.ts
@@ -33,6 +33,7 @@ import {
   transactionCommitments,
   verifyAccountUpdateSignature,
 } from '../mina-signer/src/sign-zkapp-command.js';
+import { NetworkId } from '../mina-signer/src/TSTypes.js';
 
 export {
   createTransaction,
@@ -51,6 +52,7 @@ export {
   getAccount,
   hasAccount,
   getBalance,
+  getNetworkId,
   getNetworkConstants,
   getNetworkState,
   accountCreationFee,
@@ -65,7 +67,7 @@ export {
   getProofsEnabled,
   // for internal testing only
   filterGroups,
-  type NetworkConstants
+  type NetworkConstants,
 };
 
 interface TransactionId {
@@ -374,6 +376,7 @@ interface Mina {
     tokenId?: Field
   ) => { hash: string; actions: string[][] }[];
   proofsEnabled: boolean;
+  getNetworkId(): NetworkId;
 }
 
 const defaultAccountCreationFee = 1_000_000_000;
@@ -389,12 +392,14 @@ const defaultNetworkConstants: NetworkConstants = {
 function LocalBlockchain({
   proofsEnabled = true,
   enforceTransactionLimits = true,
+  networkId = 'testnet' as NetworkId,
 } = {}) {
   const slotTime = 3 * 60 * 1000;
   const startTime = Date.now();
   const genesisTimestamp = UInt64.from(startTime);
   const ledger = Ledger.create();
   let networkState = defaultNetworkState();
+  let minaNetworkId: NetworkId = networkId;
 
   function addAccount(publicKey: PublicKey, balance: string) {
     ledger.addAccount(Ml.fromPublicKey(publicKey), balance);
@@ -421,6 +426,7 @@ function LocalBlockchain({
   > = {};
 
   return {
+    getNetworkId: () => minaNetworkId,
     proofsEnabled,
     /**
    * @deprecated use {@link Mina.getNetworkConstants}
@@ -494,7 +500,8 @@ function LocalBlockchain({
             account,
             update,
             commitments,
-            this.proofsEnabled
+            this.proofsEnabled,
+            this.getNetworkId()
           );
         }
       }
@@ -689,57 +696,63 @@ LocalBlockchain satisfies (...args: any) => Mina;
  * Represents the Mina blockchain running on a real network
  */
 function Network(graphqlEndpoint: string): Mina;
-function Network(endpoints: {
+function Network(options: {
+  networkId?: NetworkId;
   mina: string | string[];
   archive?: string | string[];
   lightnetAccountManager?: string;
 }): Mina;
 function Network(
-  input:
+  options:
     | {
+        networkId?: NetworkId;
         mina: string | string[];
         archive?: string | string[];
         lightnetAccountManager?: string;
       }
     | string
 ): Mina {
+  let minaNetworkId: NetworkId = 'testnet';
   let minaGraphqlEndpoint: string;
   let archiveEndpoint: string;
   let lightnetAccountManagerEndpoint: string;
 
-  if (input && typeof input === 'string') {
-    minaGraphqlEndpoint = input;
+  if (options && typeof options === 'string') {
+    minaGraphqlEndpoint = options;
     Fetch.setGraphqlEndpoint(minaGraphqlEndpoint);
-  } else if (input && typeof input === 'object') {
-    if (!input.mina)
+  } else if (options && typeof options === 'object') {
+    if (options.networkId) {
+      minaNetworkId = options.networkId;
+    }
+    if (!options.mina)
       throw new Error(
         "Network: malformed input. Please provide an object with 'mina' endpoint."
       );
-    if (Array.isArray(input.mina) && input.mina.length !== 0) {
-      minaGraphqlEndpoint = input.mina[0];
+    if (Array.isArray(options.mina) && options.mina.length !== 0) {
+      minaGraphqlEndpoint = options.mina[0];
       Fetch.setGraphqlEndpoint(minaGraphqlEndpoint);
-      Fetch.setMinaGraphqlFallbackEndpoints(input.mina.slice(1));
-    } else if (typeof input.mina === 'string') {
-      minaGraphqlEndpoint = input.mina;
+      Fetch.setMinaGraphqlFallbackEndpoints(options.mina.slice(1));
+    } else if (typeof options.mina === 'string') {
+      minaGraphqlEndpoint = options.mina;
       Fetch.setGraphqlEndpoint(minaGraphqlEndpoint);
     }
 
-    if (input.archive !== undefined) {
-      if (Array.isArray(input.archive) && input.archive.length !== 0) {
-        archiveEndpoint = input.archive[0];
+    if (options.archive !== undefined) {
+      if (Array.isArray(options.archive) && options.archive.length !== 0) {
+        archiveEndpoint = options.archive[0];
         Fetch.setArchiveGraphqlEndpoint(archiveEndpoint);
-        Fetch.setArchiveGraphqlFallbackEndpoints(input.archive.slice(1));
-      } else if (typeof input.archive === 'string') {
-        archiveEndpoint = input.archive;
+        Fetch.setArchiveGraphqlFallbackEndpoints(options.archive.slice(1));
+      } else if (typeof options.archive === 'string') {
+        archiveEndpoint = options.archive;
         Fetch.setArchiveGraphqlEndpoint(archiveEndpoint);
       }
     }
 
     if (
-      input.lightnetAccountManager !== undefined &&
-      typeof input.lightnetAccountManager === 'string'
+      options.lightnetAccountManager !== undefined &&
+      typeof options.lightnetAccountManager === 'string'
     ) {
-      lightnetAccountManagerEndpoint = input.lightnetAccountManager;
+      lightnetAccountManagerEndpoint = options.lightnetAccountManager;
       Fetch.setLightnetAccountManagerEndpoint(lightnetAccountManagerEndpoint);
     }
   } else {
@@ -749,6 +762,7 @@ function Network(
   }
 
   return {
+    getNetworkId: () => minaNetworkId,
     /**
    * @deprecated use {@link Mina.getNetworkConstants}
    */
@@ -1013,6 +1027,7 @@ function BerkeleyQANet(graphqlEndpoint: string) {
 }
 
 let activeInstance: Mina = {
+  getNetworkId: () => 'testnet',
   /**
    * @deprecated use {@link Mina.getNetworkConstants}
    */
@@ -1200,6 +1215,13 @@ function hasAccount(publicKey: PublicKey, tokenId?: Field): boolean {
 }
 
 /**
+ * @return The current Mina network ID.
+ */
+function getNetworkId() {
+  return activeInstance.getNetworkId();
+}
+
+/**
  * @return Data associated with the current Mina network constants.
  */
 function getNetworkConstants() {
@@ -1298,7 +1320,8 @@ async function verifyAccountUpdate(
   account: Account,
   accountUpdate: AccountUpdate,
   transactionCommitments: { commitment: bigint; fullCommitment: bigint },
-  proofsEnabled: boolean
+  proofsEnabled: boolean,
+  networkId: NetworkId
 ): Promise<void> {
   // check that that top-level updates have mayUseToken = No
   // (equivalent check exists in the Mina node)
@@ -1406,7 +1429,7 @@ async function verifyAccountUpdate(
       isValidSignature = verifyAccountUpdateSignature(
         TypesBigint.AccountUpdate.fromJSON(accountUpdateJson),
         transactionCommitments,
-        'testnet'
+        networkId
       );
     } catch (error) {
       errorTrace += '\n\n' + (error as Error).message;